+++ resources/views/layout/sidebar/fragment.blade.php
@@ -3,10 +3,6 @@
         <p class="text-lg h6 h1-dark">
             {{ ucfirst($resource->getLabel()) }}
         </p>
-
-        @if($model->fragmentModel()->exists)
-            @include('chief::layout._partials.fragment_bookmarks')
-        @endif
     </div>
 
     {!! $slot !!}
@@ -120,17 +116,17 @@
         </div>
 
         <form
-                id="detachSharedFragment{{ $model->modelReference()->get() }}"
-                method="POST"
-                action="{{ $manager->route('fragment-unshare', $owner, $model) }}"
+            id="detachSharedFragment{{ $model->modelReference()->get() }}"
+            method="POST"
+            action="{{ $manager->route('fragment-unshare', $owner, $model) }}"
         >
             @csrf
         </form>
 
         <form
-                id="changeFragmentStatus{{ $model->modelReference()->get() }}"
-                method="POST"
-                action="{{ $manager->route('fragment-status', $model) }}"
+            id="changeFragmentStatus{{ $model->modelReference()->get() }}"
+            method="POST"
+            action="{{ $manager->route('fragment-status', $model) }}"
         >
             <input type="hidden" name="online_status" value="{{ $model->fragmentModel()->isOnline()
                 ? \Thinktomorrow\Chief\Fragments\FragmentStatus::offline->value